+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 * Added/Changed/Deprecated/Removed/Fixed/Security: YOUR CHANGE HERE
+*
+* Removed: `action.can_return_nil` was removed, because it does no affect anymore
+* Removed: default `action` model was removed. Now each action must have `returns` part
 * Added: default router added. No need to assign value to constant on Router.draw
 * Added: install generator. Now it's possible to generate boilerplate code